+++ src/config.h
@@ -3,11 +3,11 @@
 
 #include <stdbool.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 
 typedef enum
 {
-	BLACK          = 30,
 	RED            = 31,
 	GREEN          = 32,
 	YELLOW         = 33,
@@ -15,7 +15,6 @@ typedef enum
 	MAGENTA        = 35,
 	CYAN           = 36,
 	WHITE          = 37,
-	BRIGHT_BLACK   = 90,
 	BRIGHT_RED     = 91,
 	BRIGHT_GREEN   = 92,
 	BRIGHT_YELLOW  = 93,
@@ -25,9 +24,6 @@ typedef enum
 	BRIGHT_WHITE   = 97
 }
 Colors_t;
-
-#define BRIGHT_RED 91
-#define RESET 0
 
 typedef struct
 {
@@ -39,9 +35,8 @@ Opt_t;
 typedef struct
 {
 	FILE* File;
-	// Opt_t Color_bar;
-	Opt_t Color_git_logo;
-	// Opt_t Color_line_number;
+	Opt_t Color_ui;
+	Opt_t Color_line_number;
 	Opt_t Color_line_number_current;
 	Opt_t Color_text;
 	Opt_t Color_warning;